# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Colegios",
        "description": "Documentación de la API (JWT Bearer)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un usuario con nombre de usuario y contraseña.",
                "operationId": "4fba3074740898952b34e5b9a9681f5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_usuario",
                                    "password"
                                ],
                                "properties": {
                                    "nombre_usuario": {
                                        "type": "string",
                                        "example": "javier01"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioSimple"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "description": "Devuelve un JWT si las credenciales son válidas.",
                "operationId": "1452175972331d50d10251ae20dfb50c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_usuario",
                                    "password"
                                ],
                                "properties": {
                                    "nombre_usuario": {
                                        "type": "string",
                                        "example": "javier01"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtener perfil del usuario autenticado",
                "operationId": "50ccc9534364638c5ddb9dc2d36ef4fd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioPerfil"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "operationId": "c463b3f92305ea956710196b78019520",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Role": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Encargado"
                    }
                },
                "type": "object"
            },
            "Persona": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "nombres": {
                        "type": "string",
                        "example": "Javier"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Salvatierra"
                    },
                    "rut": {
                        "type": "string",
                        "example": "12.345.678-9"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56 9 1234 5678",
                        "nullable": true
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "javier@example.com",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UsuarioSimple": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "persona_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "nombre_usuario": {
                        "type": "string",
                        "example": "javier01"
                    }
                },
                "type": "object"
            },
            "UsuarioPerfil": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "nombre_usuario": {
                        "type": "string",
                        "example": "javier01"
                    },
                    "persona_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "persona": {
                        "$ref": "#/components/schemas/Persona"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object"
            },
            "TokenResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGci..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Sesión cerrada correctamente"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}